finished the function
