+++ src/coordinates/Controller.java
@@ -17,6 +17,58 @@ public class Controller {
     public void initialize(){
     }
 
+    public Pane blabla(int line_length){
+        Pane pane = new Pane();
+
+        Line l1 = new Line();
+        l1.setStartX(0);
+        l1.setStartY(10);
+        l1.setEndX(line_length);
+        l1.setEndY(10);
+
+        Line l2 = new Line();
+        l2.setStartX(0);
+        l2.setStartY(20);
+        l2.setEndX(line_length);
+        l2.setEndY(20);
+
+        Line l3 = new Line();
+        l3.setStartX(0);
+        l3.setStartY(30);
+        l3.setEndX(line_length);
+        l3.setEndY(30);
+
+        Line l4 = new Line();
+        l4.setStartX(0);
+        l4.setStartY(40);
+        l4.setEndX(line_length);
+        l4.setEndY(40);
+
+        Line l5 = new Line();
+        l5.setStartX(0);
+        l5.setStartY(50);
+        l5.setEndX(line_length);
+        l5.setEndY(50);
+
+        pane.getChildren().add(l1);
+        pane.getChildren().add(l2);
+        pane.getChildren().add(l3);
+        pane.getChildren().add(l4);
+        pane.getChildren().add(l5);
+
+        javafx.scene.image.Image image = new javafx.scene.image.Image(getClass().getResource("resources/bilder_noten/Violinschluessel.png").toExternalForm());
+        ImageView notenSchlüssel = new ImageView(image);
+        pane.getChildren().add(notenSchlüssel);
+        int vio_size = 40;
+        notenSchlüssel.setFitHeight(vio_size * 1.705);
+        notenSchlüssel.setFitWidth(vio_size);
+        notenSchlüssel.setY(-5);
+        pane.setOnMouseMoved(this::onMouseMoved);
+
+        return pane;
+
+    }
+
     public void onMouseEntered(javafx.scene.input.MouseEvent mouseEvent) {
         javafx.scene.image.Image image = new javafx.scene.image.Image(getClass().getResource("note.png").toExternalForm());
         ImageView hoveredImage = (ImageView) mouseEvent.getSource();
